fix(wyh):ppg polish
