+++ ding/model/common/__init__.py
@@ -1,4 +1,4 @@
 from .head import DiscreteHead, DuelingHead, DistributionHead, RainbowHead, QRDQNHead, \
     QuantileHead, RegressionHead, ReparameterizationHead, MultiHead, head_cls_map
-from .encoder import ConvEncoder, FCEncoder, ImpalaEncoder
+from .encoder import ConvEncoder, FCEncoder, ImpalaConvEncoder
 from .utils import create_model
