+++ ding/policy/command_mode_policy_instance.py
@@ -13,7 +13,7 @@ from .r2d2_collect_traj import R2D2CollectTrajPolicy
 from .sqn import SQNPolicy
 from .ppo import PPOPolicy, PPOOffPolicy
 from .offppo_collect_traj import OffPPOCollectTrajPolicy
-from .ppg import PPGPolicy, PPGOnPolicy
+from .ppg import PPGPolicy, PPGOffPolicy
 from .a2c import A2CPolicy
 from .impala import IMPALAPolicy
 from .ngu import NGUPolicy
@@ -172,13 +172,13 @@ class IMPALACommandModePolicy(IMPALAPolicy, DummyCommandModePolicy):
     pass
 
 
-@POLICY_REGISTRY.register('ppg_command')
-class PPGCommandModePolicy(PPGPolicy, DummyCommandModePolicy):
+@POLICY_REGISTRY.register('ppg_offpolicy_command')
+class PPGOffCommandModePolicy(PPGOffPolicy, DummyCommandModePolicy):
     pass
 
 
-@POLICY_REGISTRY.register('ppg_onpolicy_command')
-class OnPPGCommandModePolicy(PPGOnPolicy, DummyCommandModePolicy):
+@POLICY_REGISTRY.register('ppg_command')
+class PPGCommandModePolicy(PPGPolicy, DummyCommandModePolicy):
     pass
