+++ dizoo/atari/config/serial/pong/pong_ppg_config.py
@@ -65,7 +65,7 @@ pong_ppg_create_config = dict(
         import_names=['dizoo.atari.envs.atari_env'],
     ),
     env_manager=dict(type='subprocess'),
-    policy=dict(type='ppg'),
+    policy=dict(type='ppg_offpolicy'),
 )
 create_config = EasyDict(pong_ppg_create_config)
