+++ dizoo/atari/config/serial/spaceinvaders/spaceinvaders_ppg_config.py
@@ -68,7 +68,7 @@ spaceinvaders_ppg_create_config = dict(
         import_names=['dizoo.atari.envs.atari_env'],
     ),
     env_manager=dict(type='subprocess'),
-    policy=dict(type='ppg'),
+    policy=dict(type='ppg_offpolicy'),
 )
 spaceinvaders_ppg_create_config = EasyDict(spaceinvaders_ppg_create_config)
 create_config = EasyDict(spaceinvaders_ppg_create_config)
