--- dizoo/procgen/bigfish/envs/__init__.py
@@ -1,1 +1,0 @@
-from .bigfish_env import BigfishEnv
