--- dizoo/procgen/coinrun/envs/__init__.py
@@ -1,1 +1,0 @@
-from .coinrun_env import CoinRunEnv
