+++ dizoo/procgen/entry/__init__.py
@@ -1,0 +1,2 @@
+from .coinrun_dqn_config import main_config, create_config
+from .coinrun_ppo_config import main_config, create_config
