+++ dizoo/procgen/envs/__init__.py
@@ -1,0 +1,1 @@
+from .procgen_env import ProcgenEnv
