--- dizoo/procgen/maze/entry/__init__.py
@@ -1,2 +1,0 @@
-from .maze_ppo_config import main_config, create_config
-from .maze_dqn_config import main_config, create_config
