--- dizoo/procgen/maze/envs/__init__.py
@@ -1,1 +1,0 @@
-from .maze_env import MazeEnv
